move some support libs from server to java
