+++ generators/java/support/doc.spec.ts
