+++ generators/server/support/prepare-relationship.ts
@@ -1,4 +1,4 @@
-import { formatDocAsApiDescription, formatDocAsJavaDoc } from './doc.js';
+import { formatDocAsApiDescription, formatDocAsJavaDoc } from '../../java/support/doc.js';
 
 export function prepareRelationship({ relationship }: { relationship: any; entity: any }) {
   if (relationship.documentation) {
